LE_Kotlin: extracted & renamed interfaces inside "publicApi" package
