+++ Attt_LocalEngine_Kotlin/src/main/kotlin/publicApi/AtttPlayer.kt
@@ -1,0 +1,12 @@
+package publicApi
+
+interface AtttPlayer {
+
+    fun getId(): Int
+
+    fun getName(): String?
+
+    fun getSymbol(): Char?
+
+    fun getMaxLineLength(): Int
+}
